mesh-1466: 1. extract RT commons
