+++ pallets/runtime/common/src/lib.rs
@@ -19,6 +19,7 @@
 
 pub mod cdd_check;
 pub mod impls;
+pub mod runtime;
 
 pub use cdd_check::CddChecker;
 pub use sp_runtime::{Perbill, Permill};
